% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinx.R
\name{ctanx}
\alias{ctanx}
\title{Show a sinx fortune when starting R}
\usage{
ctanx(method = c("add", "remove"))
}
\arguments{
\item{method}{(character) add or remove `sinx::tanx()` in ~/.Rprofile}
}
\value{
a new ~/.Rprofile
}
\description{
Show a sinx fortune when starting R
}
\examples{
ctanx()
}
