% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinx.R
\name{print.sinx}
\alias{print.sinx}
\title{Print R sinx fortunes.}
\usage{
\method{print}{sinx}(x, ...)
}
\arguments{
\item{x}{an object of class "sinx", usually a single row from `sinxs.data`.}

\item{...}{potential further arguments passed to `grep`.}
}
\value{
print.
}
\description{
Print R sinx fortunes.
}
