% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cscx.R
\name{cscx}
\alias{cscx}
\title{Create a Skeleton in a Clipboard for a new sinX}
\usage{
cscx()
}
\value{
a skeleton text for contributing to the database
}
\description{
Create a Skeleton in a Clipboard for a new sinX
}
\examples{
cscx()
}
