% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secx.R
\name{secx}
\alias{secx}
\title{Create a pagedown book from a sinx library}
\usage{
secx(lib = "tangshi", title = "A Pagedown Book", editor = "dapeng", section = "author", 
    subsection = "context", file = NULL, if_render = TRUE)
}
\arguments{
\item{lib}{library name of the sayings. See `?read.sinxs`.}

\item{title}{title of the book}

\item{editor}{editor name}

\item{section}{name of a sinx libary column, in which the characters are the sections in the book}

\item{subsection}{name of a sinx libary column, in which the characters are the subsections in the book}

\item{file}{file name of the book. Without extension.}

\item{if_render}{if render the book}
}
\value{
a book
}
\description{
Create a pagedown book from a sinx library
}
\examples{
\dontrun{
secx()

secx(lib = "sinxs", title = "cosx", subsection = "date")
}
}
